+++ buildx-web/src/api/info.ts
@@ -1,0 +1,10 @@
+import { apiFetch } from "./client";
+
+export type ServerInfo = {
+  name: string;
+  version: string;
+};
+
+export function fetchServerInfo(): Promise<ServerInfo> {
+  return apiFetch<ServerInfo>("/~api/v1/info");
+}